requirements.txt
dlt==1.9.0
pyspark>=3.5.0
delta-spark>=3.0.0
databricks-connect>=15.4
PyYAML>=6.0.0
typing-extensions>=4.0.0
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0 